Sistemata la View, assieme alle GUIViews
